chore: remove unnecessary SqlRunnerResultsTransformerFE class
