+++ packages/frontend/src/components/DataViz/store/lineChartSlice.ts
@@ -4,8 +4,8 @@ import {
     isLineChartSQLConfig,
 } from '@lightdash/common';
 import { createSlice } from '@reduxjs/toolkit';
+import { SqlRunnerResultsTransformer } from '../../../features/sqlRunner/SqlResultsRunner';
 import { setSavedChartData } from '../../../features/sqlRunner/store/sqlRunnerSlice';
-import { SqlRunnerResultsTransformerFE } from '../../../features/sqlRunner/transformers/SqlRunnerResultsTransformerFE';
 import {
     cartesianChartConfigSlice,
     onResults,
@@ -21,7 +21,7 @@ export const lineChartConfigSlice = createSlice({
         builder.addCase(onResults, (state, action) => {
             if (action.payload) {
                 const sqlRunnerResultsTransformer =
-                    new SqlRunnerResultsTransformerFE({
+                    new SqlRunnerResultsTransformer({
                         rows: action.payload.results,
                         columns: action.payload.columns,
                     });
