+++ packages/frontend/src/pages/ViewSqlChart.tsx
@@ -26,6 +26,7 @@ import { Table } from '../components/DataViz/visualizations/Table';
 import { Header } from '../features/sqlRunner/components/Header';
 import { useSavedSqlChart } from '../features/sqlRunner/hooks/useSavedSqlCharts';
 import { useSqlChartResults } from '../features/sqlRunner/hooks/useSqlChartResults';
+import { SqlRunnerResultsTransformer } from '../features/sqlRunner/SqlResultsRunner';
 import { store } from '../features/sqlRunner/store';
 import {
     useAppDispatch,
@@ -36,7 +37,6 @@ import {
     setProjectUuid,
     setSavedChartData,
 } from '../features/sqlRunner/store/sqlRunnerSlice';
-import { SqlRunnerResultsTransformerFE } from '../features/sqlRunner/transformers/SqlRunnerResultsTransformerFE';
 
 enum TabOption {
     CHART = 'chart',
@@ -84,7 +84,7 @@ const ViewSqlChart = () => {
 
     const transformer = useMemo(
         () =>
-            new SqlRunnerResultsTransformerFE({
+            new SqlRunnerResultsTransformer({
                 rows: data?.results ?? [],
                 columns: data?.columns ?? [],
             }),
